+++ dashboard/src/components/CreateActionModal.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { FormGroup, Modal, ModalBody, ModalHeader, Input, Label } from 'reactstrap';
+import { FormGroup, Input, Label } from 'reactstrap';
 import { Formik } from 'formik';
 import { toast } from 'react-toastify';
 import DatePicker from 'react-datepicker';
@@ -18,6 +18,7 @@ import useCreateReportAtDateIfNotExist from '../services/useCreateReportAtDateIf
 import { commentsState, prepareCommentForEncryption } from '../recoil/comments';
 import ActionsCategorySelect from './tailwind/ActionsCategorySelect';
 import { groupsState } from '../recoil/groups';
+import { ModalBody, ModalContainer, ModalHeader } from './tailwind/Modal';
 
 const CreateActionModal = ({ person = null, persons = null, isMulti = false, completedAt = null, dueAt, open = false, setOpen = () => {} }) => {
   const teams = useRecoilValue(teamsState);
@@ -46,9 +47,9 @@ const CreateActionModal = ({ person = null, persons = null, isMulti = false, com
   if (['restricted-access'].includes(user.role)) return null;
 
   return (
-    <Modal isOpen={open} toggle={() => setOpen(false)} size="lg" backdrop="static">
-      <ModalHeader toggle={() => setOpen(false)}>{'Créer une nouvelle action'}</ModalHeader>
-      <ModalBody>
+    <ModalContainer open={open} onClose={() => setOpen(false)} size="3xl">
+      <ModalHeader title="Créer une nouvelle action" />
+      <ModalBody className="tw-px-4 tw-py-2">
         <Formik
           initialValues={{
             name: '',
@@ -302,7 +303,7 @@ const CreateActionModal = ({ person = null, persons = null, isMulti = false, com
           }}
         </Formik>
       </ModalBody>
-    </Modal>
+    </ModalContainer>
   );
 };