+++ dashboard/src/recoil/reports.js
@@ -1,6 +1,7 @@
 import { setCacheItem } from '../services/dataManagement';
-import { atom } from 'recoil';
+import { atom, selector } from 'recoil';
 import { capture } from '../services/sentry';
+import { organisationState } from './auth';
 
 const collectionName = 'report';
 export const reportsState = atom({
@@ -20,34 +21,43 @@ export const reportsState = atom({
           }, {})
         ).filter(([key, reportsByDate]) => reportsByDate.length > 1);
         if (duplicateReports.length > 0) {
-          capture('Duplicated reports', {
-            extra: {
-              duplicateReports: duplicateReports.map(([key, reportsByDate]) => [
-                key,
-                reportsByDate.map((report) => ({
+          for (const [key, reportsByDate] of duplicateReports) {
+            capture('Duplicated reports ' + key, {
+              extra: {
+                [key]: reportsByDate.map((report) => ({
                   _id: report._id,
                   date: report.date,
                   team: report.team,
                   services: report.services,
-                  description: report.description,
-                  collaborations: report.collaborations,
-                })),
-              ]),
-              duplicateReportsIds: duplicateReports.map(([key, reportsByDate]) => [
-                key,
-                reportsByDate.map((report) => ({
-                  _id: report._id,
-                  date: report.date,
-                  team: report.team,
                   createdAt: report.createdAt,
                   deletedAt: report.deletedAt,
+                  description: report.description,
+                  collaborations: report.collaborations,
+                  organisation: report.organisation,
                 })),
-              ]),
-            },
-          });
+              },
+            });
+          }
         }
       }),
   ],
+});
+
+export const servicesSelector = selector({
+  key: 'servicesSelector',
+  get: ({ get }) => {
+    const organisation = get(organisationState);
+    if (organisation.groupedServices) return organisation.groupedServices;
+    return [{ groupTitle: 'Tous mes services', services: organisation.services ?? [] }];
+  },
+});
+
+export const flattenedServicesSelector = selector({
+  key: 'flattenedServicesSelector',
+  get: ({ get }) => {
+    const groupedServices = get(servicesSelector);
+    return groupedServices.reduce((allServices, { services }) => [...allServices, ...services], []);
+  },
 });
 
 const encryptedFields = ['description', 'services', 'team', 'date', 'collaborations', 'oldDateSystem'];