+++ dashboard/src/scenes/stats/index.js
@@ -1,5 +1,6 @@
-import React, { useCallback, useMemo, useState } from 'react';
+import React, { useCallback, useMemo } from 'react';
 import styled from 'styled-components';
+import { useLocalStorage } from 'react-use';
 import { Col, Label, Nav, NavItem, NavLink, Row, TabContent, TabPane } from 'reactstrap';
 import { useRecoilValue } from 'recoil';
 import { HeaderStyled, RefreshButton, Title as HeaderTitle } from '../../components/header';
@@ -42,7 +43,7 @@ const getDataForPeriod = (
   { startDate, endDate },
   selectedTeams,
   viewAllOrganisationData,
-  { filters = [], field = 'createdAt' } = {},
+  { filters = [], field = 'createdAt', backupField = 'createdAt' } = {},
   callback = null
 ) => {
   if (!!filters?.filter((f) => Boolean(f?.value)).length) data = filterData(data, filters);
@@ -55,7 +56,11 @@ const getDataForPeriod = (
     return callback(data, offsetHours);
   }
   return data.filter((item) =>
-    getIsDayWithinHoursOffsetOfPeriod(item[field] || item.createdAt, { referenceStartDay: startDate, referenceEndDay: endDate }, offsetHours)
+    getIsDayWithinHoursOffsetOfPeriod(
+      item[field] || item[backupField] || item.createdAt,
+      { referenceStartDay: startDate, referenceEndDay: endDate },
+      offsetHours
+    )
   );
 };
 const tabs = ['Général', 'Accueil', 'Actions', 'Personnes suivies', 'Passages', 'Rencontres', 'Observations', 'Comptes-rendus', 'Consultations'];
@@ -82,14 +87,15 @@ const Stats = () => {
   const groupsCategories = useRecoilValue(actionsCategoriesSelector);
   const { isLoading } = useDataLoader({ refreshOnMount: true });
 
-  const [selectedTerritories, setSelectedTerritories] = useState([]);
-  const [activeTab, setActiveTab] = useState(0);
-  const [filterPersons, setFilterPersons] = useState([]);
-  const [viewAllOrganisationData, setViewAllOrganisationData] = useState(teams.length === 1);
-  const [period, setPeriod] = useState({ startDate: null, endDate: null });
-  const [actionsStatuses, setActionsStatuses] = useState(DONE);
-
-  const [selectedTeams, setSelectedTeams] = useState([currentTeam]);
+  const [selectedTerritories, setSelectedTerritories] = useLocalStorage('stats-territories', []);
+  const [activeTab, setActiveTab] = useLocalStorage('stats-tab', 0);
+  const [filterPersons, setFilterPersons] = useLocalStorage('stats-filterPersons-defaultEverybody', [
+    { field: 'outOfActiveList', value: "Oui et non (c'est-à-dire tout le monde)", type: 'multi-choice' },
+  ]);
+  const [viewAllOrganisationData, setViewAllOrganisationData] = useLocalStorage('stats-viewAllOrganisationData', teams.length === 1);
+  const [period, setPeriod] = useLocalStorage('period', { startDate: null, endDate: null });
+  const [actionsStatuses, setActionsStatuses] = useLocalStorage('stats-actionsStatuses', DONE);
+  const [selectedTeams, setSelectedTeams] = useLocalStorage('stats-teams', [currentTeam]);
 
   useTitle(`${tabs[activeTab]} - Statistiques`);
 
@@ -107,7 +113,7 @@ const Stats = () => {
   const persons = useMemo(
     () =>
       getDataForPeriod(filterByTeam(allPersons, 'assignedTeams'), period, selectedTeams, viewAllOrganisationData, {
-        filters: filterPersons,
+        filters: filterPersons.filter((f) => f.field !== 'outOfActiveList'),
         field: 'followedSince',
       }),
     [allPersons, filterByTeam, filterPersons, period, selectedTeams, viewAllOrganisationData]
@@ -121,7 +127,7 @@ const Stats = () => {
         selectedTeams,
         viewAllOrganisationData,
         {
-          filters: filterPersons,
+          filters: filterPersons.filter((f) => f.field !== 'outOfActiveList'),
           field: 'followedSince',
         },
         (data, offsetHours) => {
@@ -137,6 +143,8 @@ const Stats = () => {
               item.comments?.some((a) => getIsDayWithinHoursOffsetOfPeriod(a.updatedAt, ...params)) ||
               item.passages?.some((a) => getIsDayWithinHoursOffsetOfPeriod(a.createdAt, ...params)) ||
               item.passages?.some((a) => getIsDayWithinHoursOffsetOfPeriod(a.updatedAt, ...params)) ||
+              item.rencontres?.some((a) => getIsDayWithinHoursOffsetOfPeriod(a.createdAt, ...params)) ||
+              item.rencontres?.some((a) => getIsDayWithinHoursOffsetOfPeriod(a.updatedAt, ...params)) ||
               item.relsPersonPlace?.some((a) => getIsDayWithinHoursOffsetOfPeriod(a.createdAt, ...params)) ||
               item.relsPersonPlace?.some((a) => getIsDayWithinHoursOffsetOfPeriod(a.updatedAt, ...params)) ||
               item.treatments?.some((a) => getIsDayWithinHoursOffsetOfPeriod(a.createdAt, ...params)) ||
@@ -152,15 +160,17 @@ const Stats = () => {
   );
 
   const personsForStats = useMemo(() => {
-    return filterPersons.find((f) => f.field === 'outOfActiveList' && f.value === 'Oui')
-      ? persons.filter((p) => p.outOfActiveList)
-      : persons.filter((p) => !p.outOfActiveList);
+    const outOfActiveListFilter = filterPersons.find((f) => f.field === 'outOfActiveList')?.value;
+    if (outOfActiveListFilter === 'Oui') return persons.filter((p) => p.outOfActiveList);
+    if (outOfActiveListFilter === 'Non') return persons.filter((p) => !p.outOfActiveList);
+    return persons;
   }, [filterPersons, persons]);
 
   const personsUpdatedForStats = useMemo(() => {
-    return filterPersons.find((f) => f.field === 'outOfActiveList' && f.value === 'Oui')
-      ? personsUpdated.filter((p) => p.outOfActiveList)
-      : personsUpdated.filter((p) => !p.outOfActiveList);
+    const outOfActiveListFilter = filterPersons.find((f) => f.field === 'outOfActiveList')?.value;
+    if (outOfActiveListFilter === 'Oui') return personsUpdated.filter((p) => p.outOfActiveList);
+    if (outOfActiveListFilter === 'Non') return personsUpdated.filter((p) => !p.outOfActiveList);
+    return personsUpdated;
   }, [filterPersons, personsUpdated]);
 
   const personsWithActions = useMemo(() => {
@@ -175,7 +185,11 @@ const Stats = () => {
   }, [period.endDate, period.startDate, personsUpdatedForStats, selectedTeams, viewAllOrganisationData]);
 
   const actions = useMemo(
-    () => getDataForPeriod(filterByTeam(allActions, 'team'), period, selectedTeams, viewAllOrganisationData),
+    () =>
+      getDataForPeriod(filterByTeam(allActions, 'team'), period, selectedTeams, viewAllOrganisationData, {
+        field: 'completedAt',
+        backupField: 'dueAt',
+      }),
     [allActions, filterByTeam, period, selectedTeams, viewAllOrganisationData]
   );
 
@@ -183,8 +197,8 @@ const Stats = () => {
     () => actions.filter((a) => !actionsStatuses.length || actionsStatuses.includes(a.status)),
     [actions, actionsStatuses]
   );
-  const [actionsCategoriesGroups, setActionsCategoriesGroups] = useState([]);
-  const [actionsCategories, setActionsCategories] = useState([]);
+  const [actionsCategoriesGroups, setActionsCategoriesGroups] = useLocalStorage('stats-catGroups', []);
+  const [actionsCategories, setActionsCategories] = useLocalStorage('stats-categories', []);
 
   const filterableActionsCategories = useMemo(() => {
     if (!actionsCategoriesGroups.length) return allCategories;
@@ -332,7 +346,15 @@ const Stats = () => {
 
   // Add enabled custom fields in filters.
   const filterPersonsWithAllFields = [
-    ...filterPersonsBase,
+    ...filterPersonsBase.map((f) =>
+      f.field !== 'outOfActiveList'
+        ? f
+        : {
+            ...f,
+            options: ['Oui', 'Non', "Oui et non (c'est-à-dire tout le monde)"],
+            type: 'multi-choice',
+          }
+    ),
     ...customFieldsPersonsSocial.filter((a) => a.enabled || a.enabledTeams?.includes(currentTeam._id)).map((a) => ({ field: a.name, ...a })),
     ...customFieldsPersonsMedical.filter((a) => a.enabled || a.enabledTeams?.includes(currentTeam._id)).map((a) => ({ field: a.name, ...a })),
     ...customFieldsMedicalFile.filter((a) => a.enabled || a.enabledTeams?.includes(currentTeam._id)).map((a) => ({ field: a.name, ...a })),